\name{find_elbow}
\alias{find_elbow}
\title{Detecting elbow point from density plot.}
\usage{
find_elbow(x, y)
}
\arguments{
  \item{x}{ a vector, the indices of density object. }
  \item{y}{ a vector, the density value of density object. }
}
\description{
Detecting elbow point from density plot.
}
\examples{
}
